+++ Hoa/Test/.bootstrap.atoum.php
@@ -5,6 +5,3 @@ if (isset($_SERVER['HOA_PREVIOUS_CWD'])) {
 }
 
 require_once __DIR__ . DIRECTORY_SEPARATOR . '.autoload.atoum.php';
-
-Hoa\Core::enableErrorHandler();
-Hoa\Core::enableExceptionHandler();