+++ Hoa/Test/Bin/Run.php
@@ -8,7 +8,7 @@
  *
  * New BSD License
  *
- * Copyright © 2007-2015, Hoa community. All rights reserved.
+ * Copyright © 2007-2016, Hoa community. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -36,8 +36,8 @@
 
 namespace Hoa\Test\Bin;
 
+use Hoa\Consistency;
 use Hoa\Console;
-use Hoa\Core;
 use Hoa\File;
 
 /**
@@ -45,7 +45,7 @@ use Hoa\File;
  *
  * Run tests.
  *
- * @copyright  Copyright © 2007-2015 Hoa community
+ * @copyright  Copyright © 2007-2016 Hoa community
  * @license    New BSD License
  */
 class Run extends Console\Dispatcher\Kit
@@ -97,7 +97,7 @@ class Run extends Console\Dispatcher\Kit
                     foreach ($iterator as $fileinfo) {
                         $libraryName = $fileinfo->getBasename();
 
-                        if (true === Core\Consistency::isKeyword($libraryName)) {
+                        if (true === Consistency::isKeyword($libraryName)) {
                             continue;
                         }