+++ Hoa/Test/Protocol/Vfs.php
@@ -8,7 +8,7 @@
  *
  * New BSD License
  *
- * Copyright © 2007-2015, Hoa community. All rights reserved.
+ * Copyright © 2007-2016, Hoa community. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -37,17 +37,17 @@
 namespace Hoa\Test\Protocol;
 
 use atoum;
-use Hoa\Core;
+use Hoa\Protocol;
 
 /**
  * Class \Hoa\Test\Protocol\Vfs.
  *
- * Create the hoa://Test/Vfs/ component.
+ * Create the `hoa://Test/Vfs/` node.
  *
- * @copyright  Copyright © 2007-2015 Hoa community
+ * @copyright  Copyright © 2007-2016 Hoa community
  * @license    New BSD License
  */
-class Vfs extends Core\Protocol
+class Vfs extends Protocol\Node
 {
     /**
      * Component's name.
@@ -86,7 +86,7 @@ class Vfs extends Core\Protocol
         }
 
         if (!isset($components['query'])) {
-            return 'atoum:/' . Core\Protocol::NO_RESOLUTION;
+            return 'atoum:/' . Protocol::NO_RESOLUTION;
         }
 
         parse_str($components['query'], $queries);
@@ -97,7 +97,7 @@ class Vfs extends Core\Protocol
         } elseif ('file' === $queries['type']) {
             $file = atoum\mock\streams\fs\file::get($path);
         } else {
-            return 'atoum:/' . Core\Protocol::NO_RESOLUTION;
+            return 'atoum:/' . Protocol::NO_RESOLUTION;
         }
 
         $parentDirectory = dirname($path);